facets_atlasmaker/requirements.txt
absl-py
joblib
mock
nose
pillow
pylint
requests